+++ src/Html2OpenXml/Collections/ParagraphStyleCollection.cs
@@ -44,7 +44,7 @@ namespace HtmlToOpenXml
 			{
 				TagsAtSameLevel tagsOfSameLevel = en.Current.Value.Peek();
 				foreach (OpenXmlElement tag in tagsOfSameLevel.Array)
-					SetProperties(properties, tag.CloneNode(true));
+					properties.AddChild(tag.CloneNode(true));
 			}
 		}
 
@@ -82,7 +82,7 @@ namespace HtmlToOpenXml
 			{
 				try
 				{
-#if !NET_CORE
+#if !NETSTANDARD1_3
                     var ci = System.Globalization.CultureInfo.GetCultureInfo(attrValue);
 #else
                     var ci = new System.Globalization.CultureInfo(attrValue);
@@ -142,14 +142,14 @@ namespace HtmlToOpenXml
 				if (!border.IsEmpty)
 				{
 					ParagraphBorders borders = new ParagraphBorders();
-					if (border.Top.IsValid) borders.Append(
-						new TopBorder() { Val = border.Top.Style, Color = border.Top.Color.ToHexString(), Size = (uint) border.Top.Width.ValueInPx * 4, Space = 1U });
-                    if (border.Left.IsValid) borders.Append(
-                        new LeftBorder() { Val = border.Left.Style, Color = border.Left.Color.ToHexString(), Size = (uint) border.Left.Width.ValueInPx * 4, Space = 1U });
-                    if (border.Bottom.IsValid) borders.Append(
-                        new BottomBorder() { Val = border.Bottom.Style, Color = border.Bottom.Color.ToHexString(), Size = (uint) border.Bottom.Width.ValueInPx * 4, Space = 1U });
-                    if (border.Right.IsValid) borders.Append(
-						new RightBorder() { Val = border.Right.Style, Color = border.Right.Color.ToHexString(), Size = (uint) border.Right.Width.ValueInPx * 4, Space = 1U });
+					if (border.Top.IsValid)
+						borders.TopBorder = new TopBorder() { Val = border.Top.Style, Color = border.Top.Color.ToHexString(), Size = (uint) border.Top.Width.ValueInPx * 4, Space = 1U };
+                    if (border.Left.IsValid)
+						borders.LeftBorder = new LeftBorder() { Val = border.Left.Style, Color = border.Left.Color.ToHexString(), Size = (uint) border.Left.Width.ValueInPx * 4, Space = 1U };
+                    if (border.Bottom.IsValid)
+						borders.BottomBorder = new BottomBorder() { Val = border.Bottom.Style, Color = border.Bottom.Color.ToHexString(), Size = (uint) border.Bottom.Width.ValueInPx * 4, Space = 1U };
+                    if (border.Right.IsValid)
+						borders.RightBorder = new RightBorder() { Val = border.Right.Style, Color = border.Right.Color.ToHexString(), Size = (uint) border.Right.Width.ValueInPx * 4, Space = 1U };
 
 					containerStyleAttributes.Add(borders);
 					newParagraph = true;
@@ -179,7 +179,6 @@ namespace HtmlToOpenXml
 					if (className != null)
 					{
 						containerStyleAttributes.Add(new ParagraphStyleId() { Val = className });
-                        newParagraph = true;
 						break;
 					}
 				}