+++ src/Html2OpenXml/Collections/RunStyleCollection.cs
@@ -44,7 +44,7 @@ namespace HtmlToOpenXml
 			{
 				TagsAtSameLevel tagsOfSameLevel = en.Current.Value.Peek();
 				foreach (OpenXmlElement tag in tagsOfSameLevel.Array)
-					SetProperties(properties, tag.CloneNode(true));
+					properties.AddChild(tag.CloneNode(true));
 			}
 
 			if (this.DefaultRunStyle != null)