+++ src/Html2OpenXml/PredefinedStyles.cs
@@ -1,4 +1,5 @@
-﻿using System.Resources;
+﻿using System.Reflection;
+using System.Resources;
 
 namespace HtmlToOpenXml
 {
@@ -30,8 +31,11 @@ namespace HtmlToOpenXml
                 if (object.ReferenceEquals(resourceMan, null))
                 {
                     ResourceManager temp = new ResourceManager("HtmlToOpenXml.PredefinedStyles",
-                        typeof(PredefinedStyles).Assembly);
-
+#if !NETSTANDARD1_3
+            typeof(PredefinedStyles).Assembly);
+#else
+            typeof(PredefinedStyles).GetTypeInfo().Assembly);
+#endif
                     resourceMan = temp;
                 }
                 return resourceMan;